changed bllod pressure type

+++ MH.Domain/DBModel/MedicalHistory.cs
@@ -17,7 +17,7 @@ namespace MH.Domain.DBModel
         public double? TotalColestorl { get; set; }
         public double? UricAcid { get; set; }
         public double? Lactate { get; set; }
-        public double? BloodPressue { get; set; }
+        public int? BloodPressue { get; set; }
         public string? Recomendations { get; set; }
         public int? Remidies { get; set; }
         public DateTime? NextAppointmentDate { get; set; }
